@@ -5,11 +5,11 @@
     "license": "MIT",
     "type": "project",
     "require": {
-        "php": ">=7.2",
-        "intervention/image": "^2.5",
-        "intervention/imagecache": "^2.4",
+        "php": ">=7.4",
         "laravel/lumen-framework": "^7.0",
         "league/flysystem": "^1.0",
+        "intervention/image": "^2.5",
+        "intervention/imagecache": "^2.4",
         "gabrielelana/byte-units": "^0.5.0"
     },
     "require-dev": {
