@@ -3,12 +3,13 @@
 namespace App;
 
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Storage;
 
 class Client
 {
-    public $name;
-    public $key;
+    public string $name;
+    public string $key;
 
     public function __construct(string $name, ?string $key = null)
     {
@@ -23,17 +24,21 @@ class Client
 
     public function save()
     {
-        return Storage::put($this->name . DIRECTORY_SEPARATOR . '.key', $this->key);
+        return Storage::put(
+            $this->name . DIRECTORY_SEPARATOR . '.key',
+            Hash::make($this->key),
+            'private',
+        );
     }
 
-    public static function login(string $name, string $key)
+    public static function loginOrFail(string $name, ?string $key)
     {
-        if (!Storage::exists($name . DIRECTORY_SEPARATOR . '.key')) {
-            return;
-        } elseif (Storage::get($name . DIRECTORY_SEPARATOR . '.key') !== $key) {
-            return;
+        $file = File::findOrFail($name, '.key');
+
+        if (Hash::check($key, $file->binary())) {
+            return new self($name, $key);
         }
 
-        return new self($name, $key);
+        return abort(401, 'API key is missing or invalid');
     }
 }
