@@ -6,9 +6,8 @@ use Illuminate\Support\Facades\Storage;
 
 class File
 {
-    public $name;
-    public $owner;
-    public $private;
+    public string $name;
+    public string $owner;
 
     public function __construct(string $client, string $name)
     {
@@ -51,13 +50,13 @@ class File
         return Storage::delete($this->path());
     }
 
-    public static function find(string $client, string $name)
+    public static function findOrFail(string $client, string $name)
     {
         if (Storage::exists($client . DIRECTORY_SEPARATOR . $name)) {
             return new self($client, $name);
         }
 
-        return false;
+        return abort(404, 'File not found');
     }
 
     public static function save(string $file, string $client, bool $private = false)
