@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use App\File;
 use App\Client;
 use Illuminate\Http\Request;
-use App\Http\Resources\FileResource;
+use Illuminate\Http\Response;
 use Intervention\Image\Facades\Image;
 use Laravel\Lumen\Routing\Controller as BaseController;
 
@@ -14,6 +14,7 @@ class ViewController extends BaseController
     /**
      * Files view.
      *
+     * @param Request $request
      * @param string $client
      * @param string $fileName
      *
@@ -21,53 +22,26 @@ class ViewController extends BaseController
      */
     public function view(Request $request, string $client, string $fileName)
     {
-        $file = File::find($client, $fileName);
-
-        // File not found
-        if (!$file) {
-            return abort(404);
-        }
+        $file = File::findOrFail($client, $fileName);
 
         // Private file
         if ($file->visibility() !== 'public') {
-            $request->client = Client::login($request->client, $request->header('Authorization'));
-
-            if (!$request->client) {
-                return abort(403);
-            }
+            Client::loginOrFail($request->client, $request->header('x-api-key'));
         }
 
+        // Image manipulation
         if (isset($request->w) || isset($request->h)) {
             $response = Image::cache(function ($image) use ($file, $request) {
-                return $image->make($file->binary())
-                ->resize($request->w, $request->h, function ($constraint) {
-                    $constraint->aspectRatio();
-                    $constraint->upsize();
-                });
+                return $image
+                    ->make($file->binary())
+                    ->resize($request->w, $request->h, function ($constraint) {
+                        $constraint->aspectRatio();
+                        $constraint->upsize();
+                    });
             });
         }
 
         return response($response ?? $file->binary())
             ->header('Content-Type', $file->mimeType());
-    }
-
-    /**
-     * View info about file.
-     *
-     * @param mixed $client
-     * @param mixed $fileName
-     *
-     * @return Response
-     */
-    public function info(string $client, string $fileName)
-    {
-        $file = File::find($client, $fileName);
-
-        // File not found
-        if (!$file) {
-            return abort(404);
-        }
-
-        return new FileResource($file);
     }
 }
