@@ -2,6 +2,7 @@
 
 namespace App\Http\Resources;
 
+use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
 
 class FileResource extends JsonResource
@@ -9,7 +10,7 @@ class FileResource extends JsonResource
     /**
      * Transform the resource into an array.
      *
-     * @param mixed $request
+     * @param Request $request
      *
      * @return array
      */
